Minonr updates

+++ pyradi/_cleanClutter.py
@@ -1,32 +1,56 @@
 #clean a precisely defined set of files from a precisely defined directory set.
 #the specified files are deleted, after the user has been prompted
 
-
+import os
 import os.path, fnmatch
+import sys
 
 #lists the files in a directory and subdirectories (from Python Cookbook)
 def listFiles(root, patterns='*', recurse=1, return_folders=0):
     # Expand patterns from semicolon-separated string to list
     pattern_list = patterns.split(';')
-    # Collect input and output arguments into one bunch
-    class Bunch:
-        def __init__(self, **kwds): self.__dict__.update(kwds)
-    arg = Bunch(recurse=recurse, pattern_list=pattern_list,
-        return_folders=return_folders, results=[])
+    filenames = []
+    filertn = []
 
-    def visit(arg, dirname, files):
-        # Append to arg.results all relevant files (and perhaps folders)
-        for name in files:
-            fullname = os.path.normpath(os.path.join(dirname, name))
-            if arg.return_folders or os.path.isfile(fullname):
-                for pattern in arg.pattern_list:
+    if sys.version_info[0] < 3:
+
+        # Collect input and output arguments into one bunch
+        class Bunch(object):
+            def __init__(self, **kwds): self.__dict__.update(kwds)
+        arg = Bunch(recurse=recurse, pattern_list=pattern_list,
+            return_folders=return_folders, results=[])
+
+        def visit(arg, dirname, files):
+            # Append to arg.results all relevant files (and perhaps folders)
+            for name in files:
+                fullname = os.path.normpath(os.path.join(dirname, name))
+                if arg.return_folders or os.path.isfile(fullname):
+                    for pattern in arg.pattern_list:
+                        if fnmatch.fnmatch(name, pattern):
+                            arg.results.append(fullname)
+                            break
+            # Block recursion if recursion was disallowed
+            if not arg.recurse: files[:]=[]
+        os.path.walk(root, visit, arg)
+        return arg.results
+
+    else:
+        for dirpath,dirnames,files in os.walk(root):
+            if dirpath==root or recurse:
+                for filen in files:
+                    filenames.append(os.path.abspath(os.path.join(os.getcwd(),dirpath,filen)))
+                if return_folders:
+                    for dirn in dirnames:
+                        filenames.append(os.path.abspath(os.path.join(os.getcwd(),dirpath,dirn)))
+        for name in filenames:
+            if return_folders or os.path.isfile(name):
+                for pattern in pattern_list:
                     if fnmatch.fnmatch(name, pattern):
-                        arg.results.append(fullname)
+                        filertn.append(name)
                         break
-        # Block recursion if recursion was disallowed
-        if not arg.recurse: files[:]=[]
-    os.path.walk(root, visit, arg)
-    return arg.results
+
+    return filertn
+
 
 #function to delete files in specified directory
 #  first parameter defines if the search must be recursively 0=not, 1=recursive
@@ -35,20 +59,19 @@ def listFiles(root, patterns='*', recurse=1, return_folders=0):
 #  the user is promted before the files are deleted
 def QueryDelete(recurse,dir,patn):
     thefiles = listFiles(dir, patn,recurse)
-    if len(thefiles)>0:
-        for filename in thefiles:
-            print filename
-        instr=raw_input("Delete these files? (y/n)")
-        if instr=='y':
+    if thefiles is not None:
+        if len(thefiles)>0:
             for filename in thefiles:
-                os.remove(filename)
-
-
+                print(filename)
+            instr=input("Delete these files? (y/n)")
+            if instr=='y':
+                for filename in thefiles:
+                    os.remove(filename)
 
 #we take the conservative approach and do not do blanket erase,
 #rather do it by type, asking the user first
 QueryDelete(0,'.', '*.eps;*.png;*.jpg;*.pdf;*.txt;*.tiff;*.dat;*.lut')
-QueryDelete(0,'.', 'tape7-*.txt;arr*.txt;Traje*.txt;trian*.txt;vertex*.txt')
+QueryDelete(0,'.', 'tape7-*.txt;arr*.txt;Traje*.txt;trian*.txt;vertex*.txt;*.testing')
 QueryDelete(0,'.', 'arr*.txt;colourcoordinates.*;tar;*.svg' )
 QueryDelete(0,'.', '*.hdf5')
